Refactor generator-name-and-prompts test
